The great connection string migration
