+++ src/commands/appSettings/connectionSettings/azureWebJobsStorage/AzureWebJobsStorageExecuteStep.ts
@@ -7,7 +7,7 @@ import { IStorageAccountWizardContext } from '@microsoft/vscode-azext-azureutils
 import { AzureWizardExecuteStep } from '@microsoft/vscode-azext-utils';
 import { ConnectionKey, ConnectionType, localStorageEmulatorConnectionString } from '../../../../constants';
 import { MismatchBehavior, setLocalAppSetting } from '../../../../funcConfig/local.settings';
-import { getStorageConnectionString } from '../../../../utils/azure';
+import { getStorageConnectionString } from '../getLocalConnectionSetting';
 import { IAzureWebJobsStorageWizardContext } from './IAzureWebJobsStorageWizardContext';
 
 // Todo in next PR: Refactor and inherit use from SetConnectionSettingBaseStep & remove _setConnectionForDeploy
